OperatorTakeLast
